Extract fetch results to independent function and move to client.go file
